+++ src-admin/src/Components.js
@@ -1,4 +1,4 @@
-import CatalogDevicesComponent, { LicenseActivationComponent, SettingsCsvComponent, ProviderSelectComponent, ModelSelectComponent, UsageResetComponent } from './Components.jsx';
+import CatalogDevicesComponent, { HowToComponent, LicenseActivationComponent, SettingsCsvComponent, ProviderSelectComponent, ModelSelectComponent, UsageResetComponent } from './Components.jsx';
 
-export { CatalogDevicesComponent, LicenseActivationComponent, SettingsCsvComponent, ProviderSelectComponent, ModelSelectComponent, UsageResetComponent };
-export default { CatalogDevicesComponent, LicenseActivationComponent, SettingsCsvComponent, ProviderSelectComponent, ModelSelectComponent, UsageResetComponent };
+export { CatalogDevicesComponent, HowToComponent, LicenseActivationComponent, SettingsCsvComponent, ProviderSelectComponent, ModelSelectComponent, UsageResetComponent };
+export default { CatalogDevicesComponent, HowToComponent, LicenseActivationComponent, SettingsCsvComponent, ProviderSelectComponent, ModelSelectComponent, UsageResetComponent };